+++ src/constant/GOOGLE_CONSTANT.ts
@@ -12,9 +12,9 @@ const GOOGLE_CONSTANT = {
     'accept-language': 'en-US,en',
     referer: 'https://www.google.com/',
     'upgrade-insecure-requests': '1',
-    'user-agent':
-      'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.86 Safari/537.36'
+    'User-Agent':
+      'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36'
   }
-};
+} as const;
 
 export default GOOGLE_CONSTANT;